Finished some functionality, still working on some methods. Messed a bit around with the packet system. But at the end stayed with the system as it is now.

+++ shared/src/main/java/de/datasec/phenix/shared/packetsystem/Packet.java
@@ -9,7 +9,7 @@ import java.io.*;
  */
 public abstract class Packet {
 
-    protected int id;
+    protected byte id;
 
     public abstract void read(ByteBuf byteBuf);
